vk_meshview: Add a couple of C test programs for testing vert dedup perf
The exporter script in Blender is extremely slow to export so I added a couple of
C test programs that do the equivalent work (deduplicating vertex buffer) in order
to see whether it's an algorithm issue or a Python-is-slow issue.

The result is that it's actually both, but the algorithm can be greatly improved.
The current naive implementation is a linear search through the new vertex buffer,
where the number increases steadily and results in an O(n * m) algorithm.

The improvement is to use a hash table where the key is the position values,
in order to have a significantly smaller number of verts to look through.
Even with a not-particularly-great hash function and table implementation,
the results are staggeringly improved:

	Linear Search: 5.36s
	Hashtable:     0.06s

CPU: Intel Core i5-8250U @ 1.60GHz

The next step is to implement a similar algorithm in Python and hope
that it's enough to make the exporter have a usable speed.
Otherwise it may be time to roll up the sleeves and try out Cython,
or even Blender's to_c_pointer() function on many data structures
where we can directly access the memory and thus pass it onto a C library.