vk_meshview: Implement hashtable lookup for verts, improve export speed significantly
Using a hashtable (simply just Python dict where the key is the vert)
made the performance increase infinitely and it is now acceptable.
Of course, this is still Python so the speed is not quite what it could be,
but this will do for now and for the scope of this simple exporter
that is only really meant to serve as a way to get mesh data into
these graphics examples.

Only issue is that now there are slightly less vertices deduplicated than before,
this could maybe be because we are using floats in tuples directly now
instead of the Blender vector classes. Could it be that the equality operator
in those classes used a fuzzy match with threshold instead of just direct float value equality?
In which case we should explore using a threshold.