vk_meshview: Fix incorrect memory size causing staging buffer copies to overrun buffers
Previously VkMemoryRequirements::size was stored as the buffer size for VK_Buffer,
but this is incorrect because the memory size needed for allocation can be larger.
Later on using this in the staging buffer queue caused buffer overruns.

Since the information needed for VK_Buffer is not the memory allocation size but rather
the size of the data stored within the buffer, it was changed to use the size passed in.
This fixes the validation errors and buffer overrun problem.